um: remove variable stack array in os_rcv_fd_msg()
When generalizing this, I was in the mindset of this being
"userspace" code, but even there we should not use variable
arrays as the kernel is moving away from allowing that.

Simply reserve (but not use) enough space for the maximum
two descriptors we might need now, and return an error if
attempting to receive more than that.

Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202407041459.3SYg4TEi-lkp@intel.com/
Link: https://patch.msgid.link/20240704122035.789b679b263c.I17235f263887dfd4bde6c5d5f132f974bd2f756f@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>